Adding Doorbell Service for Nest Hello
This is an attempt to add the Doorbell Service for the Nest Hello based on this.device.last_event.has_person.
I'm not a coder so if anyone can help out that can be great. 
If "this.device.last_event.has_person" can trigger a Doorbell Accessory in HomeKit reliably, in conjunction with KhaosT's homebridge-nest-cam plugin, then the Nest Hello Doorbell can function as a Doorbell and Motion Sensor - The video part of Video Doorbell would be handled by homebridge-nest-cam (as long as the two Accessories are placed in the same HomeKit Room).