Move `eslint` for plugin `dependencies` to `devDependencies`
This updates `homebridge-nest`'s dependencies so that `eslint`
is one of the `devDependencies`, rather than one of the main
`dependencies`.

This reduces the amount to install and thus the install time
when installing the plugin - especially given the `eslint`
itself has quite a lot of onward dependencies.

This should be okay, since the project does not use ESLint
at runtime. As the [`eslint` docs][1] put it:

> Make sure your plugins (and ESLint) are both in your project's package.json as devDependencies (or dependencies, if your project uses ESLint at runtime).

or in the words of [Allesandro Segala][2]:

> There are some clear-cut cases, or almost. Packages like eslint are always a devDependency …unless, of course, you’re building a CLI whose job is running eslint, in which case you’d add it as a dependency!

[1]: https://www.npmjs.com/package/eslint
[2]: https://withblue.ink/2020/06/07/is-this-a-dependency-or-devdependency.html